+++ tests/testthat/test-helpers.R
@@ -8,3 +8,18 @@ test_that("nrow() works", {
     connection_clean_up(conn)
   }
 })
+
+
+test_that("unique_table_name() works", {
+  # Store options before tests and reset (tests modify options)
+  opts <- options("SCDB_table_name" = NULL, "test_table_name" = NULL)
+
+  expect_equal(unique_table_name(), "SCDB_001")
+  expect_equal(unique_table_name(), "SCDB_002")
+
+  expect_equal(unique_table_name("test"), "test_001")
+  expect_equal(unique_table_name("test"), "test_002")
+
+  # Reset options
+  options(opts)
+})